move get_user_daily_balance_history dashboard util function to new util file and class
